+++ surfactant/database_manager/external_db_config.py
@@ -8,8 +8,10 @@
 # SPDX-License-Identifier: MIT
 # surfactant/database_manager/external_db_config.py
 import logging
-
 import tomlkit
+
+
+from surfactant.database_manager.utils import download_content  # pylint: disable=import-outside-toplevel
 
 # URL for the hosted external TOML file on ReadTheDocs
 DEFAULT_EXTERNAL_DB_CONFIG_URL = (
@@ -22,8 +24,6 @@ def fetch_external_db_config(url: str = DEFAULT_EXTERNAL_DB_CONFIG_URL) -> dict:
     Download and parse the external TOML file containing database source overrides.
     Returns an empty dict on failure.
     """
-    from .database_utils import download_content  # pylint: disable=import-outside-toplevel
-
     content = download_content(url)
 
     try:
